requirements.txt
Flask
Flask-Cors
PyJWT
python-dotenv
mysql.connector